[fix/#59]: saveTimer
